requirements.txt
tensorflow==2.3.0
pillow==6.2.0
numpy==1.16.5
streamlit
wget
